+++ src/Context.php
@@ -170,7 +170,7 @@ class Context implements ContextInterface
         $this->getBlockCloser()->closeUnmatchedBlocks();
         $block->setStartLine($this->lineNumber);
         while (!$this->tip->canContain($block)) {
-            $this->tip->finalize($this);
+            $this->tip->finalize($this, $this->lineNumber);
         }
 
         $this->tip->appendChild($block);